move replace_with_mask to utils
