+++ tests/nested_pandas/series/test_ext_array.py
@@ -12,7 +12,8 @@ from pandas.testing import assert_frame_equal, assert_series_equal
 from nested_pandas import NestedDtype
 from nested_pandas.datasets import generate_data
 from nested_pandas.nestedframe.core import NestedFrame
-from nested_pandas.series.ext_array import NestedExtensionArray, convert_df_to_pa_scalar, replace_with_mask
+from nested_pandas.series.ext_array import NestedExtensionArray, convert_df_to_pa_scalar
+from nested_pandas.series.utils import replace_with_mask
 
 
 def test_replace_with_mask():
